+++ examples/agent_horde/config/config.exs
@@ -1,0 +1,5 @@
+import Config
+
+config :normandy, adapter: Poison
+
+import_config "#{config_env()}.exs"