+++ examples/agent_horde/test/test_helper.exs
@@ -1,0 +1,4 @@
+ExUnit.configure(exclude: [:live])
+ExUnit.start()
+
+Code.require_file("support/stub_tool.exs", __DIR__)